requirements.txt
aiohttp==3.13.2
fastapi==0.128.0
mcp==1.25.0
protobuf==6.33.2
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.40.0
